Refactored how classids work.
When the instance loader finds a class, it puts the originally
requested classid in a '__classid' parameter that is sent to the
object's constructor. Calls to class_id() and in some cases name()
return this value. The old Loader::class_id() method is now deprecated,
and in fact removed entirely from the Instance class loader.

The Instance Loader now supports nested namespaces by using a dot
syntax, such as module.submodule which would turn into
\root\namespace\module\submodule when searching for the class.